Reinstating the Win64 platform and making a number of changes to the library to make it easier to work with, especially when dealing with cursors and coming from the old BDB ways.

+++ lmdbjni/src/main/java/org/fusesource/lmdbjni/Cursor.java
@@ -29,12 +29,15 @@ import static org.fusesource.lmdbjni.Util.checkErrorCode;
  * @author <a href="http://hiramchirino.com">Hiram Chirino</a>
  */
 public class Cursor extends NativeObject implements Closeable {
-
     private final Env env;
+    private final Transaction tx;
+    private final Database db;
 
-    Cursor(Env env, long self) {
+    Cursor(Env env, long self, Transaction tx, Database db) {
         super(self);
         this.env = env;
+        this.tx = tx;
+        this.db = db;
     }
 
     public void close() {
@@ -48,7 +51,7 @@ public class Cursor extends NativeObject implements Closeable {
         checkErrorCode(mdb_cursor_renew(tx.pointer(), pointer()));
     }
 
-    public Entry get(GetOp op) {
+    public Entry get(CursorOp op) {
         checkArgNotNull(op, "op");
 
         Value key = new Value();
@@ -61,8 +64,7 @@ public class Cursor extends NativeObject implements Closeable {
         return new Entry(key.toByteArray(), value.toByteArray());
     }
 
-    public Entry seek(SeekOp op, byte[] key) {
-        checkArgNotNull(key, "key");
+    public Entry get(CursorOp op, byte[] key) {
         checkArgNotNull(op, "op");
         NativeBuffer keyBuffer = NativeBuffer.create(key);
         try {
@@ -75,24 +77,73 @@ public class Cursor extends NativeObject implements Closeable {
             checkErrorCode(rc);
             return new Entry(keyValue.toByteArray(), value.toByteArray());
         } finally {
-            keyBuffer.delete();
+        	if (keyBuffer != null)
+        		keyBuffer.delete();
         }
+    }
 
+    public Entry get(CursorOp op, byte[] key, byte[] value) {
+        checkArgNotNull(op, "op");
+        NativeBuffer keyBuffer = NativeBuffer.create(key);
+        NativeBuffer valBuffer = NativeBuffer.create(value);
+        
+        try {
+            Value keyValue = keyBuffer != null ? new Value(keyBuffer) : new Value();
+            Value valValue = valBuffer != null ? new Value(valBuffer) : new Value();
+            int rc = mdb_cursor_get(pointer(), keyValue, valValue, op.getValue());
+            if( rc == MDB_NOTFOUND ) {
+                return null;
+            }
+            checkErrorCode(rc);
+            return new Entry(keyValue.toByteArray(), valValue.toByteArray());
+        } finally {
+        	if (keyBuffer != null)
+        		keyBuffer.delete();
+        	if (valBuffer != null)
+        		valBuffer.delete();
+        }
+    }
+
+    public OperationStatus get(CursorOp op, DatabaseEntry key, DatabaseEntry value) {
+        checkArgNotNull(op, "op");
+        NativeBuffer keyBuffer = NativeBuffer.create(key.getData());
+        NativeBuffer valBuffer = NativeBuffer.create(value.getData());
+        
+        try {
+            Value keyValue = keyBuffer != null ? new Value(keyBuffer) : new Value();
+            Value valValue = valBuffer != null ? new Value(valBuffer) : new Value();
+            int rc = mdb_cursor_get(pointer(), keyValue, valValue, op.getValue());
+            if (rc == MDB_NOTFOUND) {
+                return OperationStatus.NOTFOUND;
+            }
+            checkErrorCode(rc);
+            key.setData(keyValue.toByteArray());
+            value.setData(valValue.toByteArray());
+            return OperationStatus.SUCCESS;
+        } finally {
+        	if (keyBuffer != null)
+        		keyBuffer.delete();
+        	if (valBuffer != null)
+        		valBuffer.delete();
+        }
     }
 
     public byte[] put(byte[] key, byte[] value, int flags) {
         checkArgNotNull(key, "key");
         checkArgNotNull(value, "value");
         NativeBuffer keyBuffer = NativeBuffer.create(key);
+
         try {
-            NativeBuffer valueBuffer = NativeBuffer.create(value);
+            NativeBuffer valBuffer = NativeBuffer.create(value);
             try {
-                return put(keyBuffer, valueBuffer, flags);
+                return put(keyBuffer, valBuffer, flags);
             } finally {
-                valueBuffer.delete();
+            	if (valBuffer != null)
+            		valBuffer.delete();
             }
         } finally {
-            keyBuffer.delete();
+        	if (keyBuffer != null)
+        		keyBuffer.delete();
         }
     }
 
@@ -117,5 +168,14 @@ public class Cursor extends NativeObject implements Closeable {
         checkErrorCode(mdb_cursor_count(pointer(), rc));
         return rc[0];
     }
+    
+    public Database getDatabase() {
+//        long dbi = mdb_cursor_dbi(pointer());
+        return db;
+    }
 
+    public Transaction getTransaction() {
+//      long txn = mdb_cursor_txn(pointer());
+      return tx;
+  }
 }
